Yellow-main/YELLOW/requirements.txt
python-telegram-bot
python-dotenv
httpx
supabase
telethon==1.36.0
pandas